modify db schema and remove conversion

+++ core/src/server/web_impl/handler/WebRequestHandler.cpp
@@ -765,7 +765,7 @@ WebRequestHandler::Search(const std::string& collection_name, const nlohmann::js
                 nlohmann::json one_entity_json;
                 for (const auto& field : field_mappings) {
                     auto field_name = field.first->GetName();
-                    switch ((int64_t)field.first->GetFtype()) {
+                    switch (field.first->GetFtype()) {
                         case engine::DataType::INT32: {
                             int32_t int32_value;
                             int64_t offset = (i * step + j) * sizeof(int32_t);
